requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
langchain==0.0.340
openai==1.3.5
protobuf==4.25.1
python-dotenv==1.0.0
sentence_transformers==2.2.2
